[fix] Various issues to improve playability
closes https://github.com/TimPoliquin/unity-olympia-festival-of-the-gods/issues/220
closes https://github.com/TimPoliquin/unity-olympia-festival-of-the-gods/issues/221
closes https://github.com/TimPoliquin/unity-olympia-festival-of-the-gods/issues/222

Improvements to the options panel, ritual overlay highlights, and json file parsing!
