--- Assets/AttentionUI.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 987ff0d4a9bac4b1ba9d80aee120a197
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
