+++ Assets/Scripts/Animation/AttentionUI.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 818c784fdf5084b97b805fccf6b95ffb
+guid: 987ff0d4a9bac4b1ba9d80aee120a197
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
